Components structure refactored;
