+++ src/components/Releases.js
@@ -1,0 +1,21 @@
+import { ReleaseContainer } from "../styles/components/stylesReleases";
+import MovieCard from "../components/layout/MovieCard";
+
+function Releases() {
+  return (
+    <ReleaseContainer>
+      <MovieCard />
+      <MovieCard />
+      <MovieCard />
+      <MovieCard />
+      <MovieCard />
+      <MovieCard />
+      <MovieCard />
+      <MovieCard />
+      <MovieCard />
+      <MovieCard />
+    </ReleaseContainer>
+  );
+}
+
+export default Releases;
