+++ src/pages/index.js
@@ -1,17 +1,12 @@
 import Head from "next/head";
-import { useState } from "react";
-
-import SearchBox from "../components/SearchBox";
-import MovieCard from "../components/MovieCard";
 
 import { Container } from "../styles/pages/stylesHome";
-import Pagination from "@material-ui/lab/Pagination";
+
+import SearchBox from "../components/SearchBox";
+import Paginations from "../components/Paginations";
+import Releases from "../components/Releases";
 
 export default function Home() {
-  const [page, setPage] = useState(null);
-
-  const handleChangePage = (event, thisPage) => event && setPage(thisPage);
-
   return (
     <>
       <Head>
@@ -22,28 +17,10 @@ export default function Home() {
       <Container>
         <section>
           <p>Lançamentos</p>
-          <div>
-            <MovieCard />
-            <MovieCard />
-            <MovieCard />
-            <MovieCard />
-            <MovieCard />
-            <MovieCard />
-            <MovieCard />
-            <MovieCard />
-            <MovieCard />
-            <MovieCard />
-          </div>
+          <Releases />
         </section>
-        <div className="pagination">
-          <Pagination
-            count={1000}
-            onChange={handleChangePage}
-            color="secondary"
-            variant="outlined"
-            hidePrevButton
-            hideNextButton
-          />
+        <div>
+          <Paginations />
         </div>
       </Container>
     </>
