+++ src/styles/components/layout/stylesMovieCard.js
