+++ src/styles/components/stylesPagination.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Pages = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  padding-bottom: 1.6rem;
+`;
