+++ src/styles/components/stylesReleases.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const ReleaseContainer = styled.div`
+  display: grid;
+  grid-template-columns: repeat(5, 1fr);
+  gap: 1rem;
+`;
